Update searchbar to use rfh's watch over onChange
Having both onChange and onSubmit caused issues for the onSubmit to run
You were unable to use the enter key after having typed in your search
term, unless you refocused the input.

The solution then was to use react form hooks' `watch` function to watch
the input value.

I also had to move the update function out of the searchBar as this
was being rerendered and caused the debounce to not work properly.