Add header with mobile and desktop sub components
The SiteHeader checks if a media query has been hit and then displays
either mobile or desktop version of the header components. The reasoning
for doing it like this is that the layouts were vastly different, and
mobile had extra functionality that would be more work to get rid of on
desktop.

To make updates easier in the future the components are split into
smaller shared components.